requirements.txt
requests
python-dotenv
pytz
cloudscraper
apscheduler
supabase
pyTelegramBotAPI
cerebras_cloud_sdk
pandas
numpy
aiohttp
loguru
pydantic
tenacity
ratelimit
PyYAML
flask
gunicorn